move color_table to test_plot as it uses cleopatra package
