+++ tests/dataset/test_plot.py
@@ -1,5 +1,7 @@
 import pytest
 import numpy as np
+import pandas as pd
+from pandas import DataFrame
 from osgeo import gdal
 from pyramids.dataset import Dataset, Datacube
 
@@ -47,3 +49,47 @@ class TestPlotDataCube:
         cube.open_datacube()
         cleo = cube.plot()
         assert isinstance(cleo, Array)
+
+
+class TestColorTable:
+    def test_get_color_table(self, src_with_color_table: Dataset):
+        dataset = Dataset(src_with_color_table)
+        df = dataset._get_color_table()
+        assert isinstance(df, DataFrame)
+        assert all(df.columns == ["band", "values", "red", "green", "blue", "alpha"])
+        assert all(df.band == 1)
+        # test the color_table property
+        df = dataset.color_table
+        assert isinstance(df, DataFrame)
+        assert all(df.columns == ["band", "values", "red", "green", "blue", "alpha"])
+        assert all(df.band == 1)
+
+    def test_set_color_table(self, src_without_color_table: Dataset):
+        color_hex = ["#709959", "#F2EEA2", "#F2CE85", "#C28C7C", "#D6C19C"]
+        values = [1, 3, 5, 7, 9]
+        df = pd.DataFrame(columns=["band", "values", "color"])
+        df.loc[:, "values"] = values
+        df.loc[:, "band"] = 1
+        df.loc[:, "color"] = color_hex
+
+        dataset = Dataset(src_without_color_table)
+        dataset._set_color_table(df, overwrite=True)
+
+        color_table = dataset.raster.GetRasterBand(1).GetColorTable()
+        assert color_table is not None, "the color table should not be None"
+        assert color_table.GetCount() == 10, "the color table should have 5 colors"
+        colors = [color_table.GetColorEntry(i) for i in range(color_table.GetCount())]
+        assert colors == [
+            (0, 0, 0, 0),
+            (112, 153, 89, 255),
+            (0, 0, 0, 0),
+            (242, 238, 162, 255),
+            (0, 0, 0, 0),
+            (242, 206, 133, 255),
+            (0, 0, 0, 0),
+            (194, 140, 124, 255),
+            (0, 0, 0, 0),
+            (214, 193, 156, 255),
+        ]
+        # test the color_table property
+        dataset.color_table = df
